rename event_bus to eventbus
