+++ internal/provider/provider.go
@@ -33,7 +33,7 @@ import (
 	"github.com/sacloud/terraform-provider-sakura/internal/service/bridge"
 	"github.com/sacloud/terraform-provider-sakura/internal/service/container_registry"
 	"github.com/sacloud/terraform-provider-sakura/internal/service/disk"
-	"github.com/sacloud/terraform-provider-sakura/internal/service/event_bus"
+	"github.com/sacloud/terraform-provider-sakura/internal/service/eventbus"
 	"github.com/sacloud/terraform-provider-sakura/internal/service/icon"
 	"github.com/sacloud/terraform-provider-sakura/internal/service/internet"
 	"github.com/sacloud/terraform-provider-sakura/internal/service/kms"
@@ -219,8 +219,8 @@ func (p *sakuraProvider) DataSources(_ context.Context) []func() datasource.Data
 		archive.NewArchiveDataSource,
 		bridge.NewBridgeDataSource,
 		container_registry.NewContainerRegistryDataSource,
-		event_bus.NewEventBusProcessConfigurationDataSource,
-		event_bus.NewEventBusScheduleDataSource,
+		eventbus.NewEventBusProcessConfigurationDataSource,
+		eventbus.NewEventBusScheduleDataSource,
 		disk.NewDiskDataSource,
 		icon.NewIconDataSource,
 		internet.NewInternetDataSource,
@@ -245,8 +245,8 @@ func (p *sakuraProvider) Resources(_ context.Context) []func() resource.Resource
 		bridge.NewBridgeResource,
 		container_registry.NewContainerRegistryResource,
 		disk.NewDiskResource,
-		event_bus.NewEventBusProcessConfigurationResource,
-		event_bus.NewEventBusScheduleResource,
+		eventbus.NewEventBusProcessConfigurationResource,
+		eventbus.NewEventBusScheduleResource,
 		icon.NewIconResource,
 		internet.NewInternetResource,
 		kms.NewKMSResource,
