+++ internal/service/eventbus/model_schedule.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package event_bus
+package eventbus
 
 import (
 	"strconv"
